requirements.txt
annotated-types==0.7.0
anyio==4.6.0
apiron==5.1.0
bcrypt==4.2.0
beren==0.7.0
caldav==1.3.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
configparser==7.1.0
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
dicom2nifti==2.5.0
einops==0.4.1
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.9.0
Genshi==0.7.9
h11==0.14.0
icalendar==5.0.13
idna==3.8
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
monai==1.3.2
mpmath==1.3.0
networkx==3.3
nibabel==3.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
passlib==1.7.4
pendulum==3.0.0
pillow==10.4.0
pip==24.2
polib==1.2.0
progressbar==2.2
protobuf==5.28.2
psycopg2-binary==2.9.9
pycountry==20.7.3
pydantic==2.9.2
pydantic_core==2.23.4
pydicom==2.0.0
pyparsing==3.1.4
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
python-magic==0.4.27
python-multipart==0.0.12
python-sql==1.5.1
python-stdnum==1.20
pytz==2024.1
PyWebDAV3-GNUHealth==0.12.0
qrcode==7.4.2
recurring-ical-events==3.2.0
relatorio==0.10.2
requests==2.32.3
scipy==1.14.1
setuptools==75.1.0
simpleeval==0.9.13
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tensorboardX==2.1
time-machine==2.15.0
torch==2.4.1
tqdm==4.59.0
triton==3.0.0
trytond==6.0.51
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
unoconv==0.9.0
urllib3==2.2.2
uvicorn==0.31.0
vobject==0.9.7
wheel==0.44.0
wrapt==1.16.0
x-wr-timezone==0.0.7
